Use CODEOWNERS instead of dependabot reviewers (#43)
Dependabot is now posting this comment as a warning:

The reviewers field in the dependabot.yml file will be removed
soon. Please use the code owners file to specify reviewers for Dependabot
PRs. For more information, see this blog post.

https://github.blog/changelog/2025-04-29-dependabot-reviewers-configuration-option-being-replaced-by-code-owners/

CODEOWNERS is already correct in this repository, so this change only
removes the reviewers entry.